refactor: работа с лайками/дизлайками для отзывов вынесена в отдельный Storage.
